+++ client/main.js
@@ -1,4 +1,5 @@
 const createElement = require('./utils/createElement.js')
+const createFormTable = require('./utils/createFormTable.js')
 const HashRouter = require('./utils/hashRouter.js')
 
 function sendUrlPostRequest(urlSubmission) {
@@ -85,30 +86,6 @@ function submitTracks(form, numOfTracks) {
     tracklist.push(track)
   }
   return tracklist
-}
-
-function createFormTable() {
-  const $formTable =
-  createElement('row', {}, '', [
-    createElement('div', {class: 'col-md-8 col-md-offset-2 view hidden', id: 'create-tracklist'}, '', [
-      createElement('h3', {id: 'youtube-video-title'}, '', []),
-      createElement('form', {id: 'tracklist-form'}, '', [
-        createElement('table', {class: 'table table-bordered'}, '', [
-          createElement('thead', {}, '', [
-            createElement('tr', {}, '', [
-              createElement('th', {}, 'Num', []),
-              createElement('th', {}, 'Name', []),
-              createElement('th', {}, 'Start', []),
-              createElement('th', {}, 'End', [])
-            ])
-          ]),
-          createElement('tbody', {id: 'track-form-container'}, '', [])
-        ]),
-        createElement('input', {id: 'track-form-submit-button', type: 'submit', value: 'Start Slicing'}, '', [])
-      ])
-    ])
-  ])
-  return $formTable
 }
 
 function addTrackForm() {
